rename src/lib/core/app.ts to src/lib/core/WacLdp.ts
