+++ test/unit/server.test.ts
@@ -1,6 +1,5 @@
-import * as http from 'http'
-jest.mock('../../src/lib/core/app')
-import { makeHandler } from '../../src/lib/core/app'
+jest.mock('../../src/lib/core/WacLdp')
+import { makeHandler } from '../../src/lib/core/WacLdp'
 import { closeServer } from '../../src/server'
 
 test('server', () => {
